Optimize all proxy-related apis
