+++ cmd/lambda/TCR/add_proxy/config/BUILD.bazel
@@ -1,4 +1,4 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
@@ -9,12 +9,6 @@ go_library(
         "//internal/pkg/error_config:go_default_library",
         "//internal/platform/postgres_config/TCR/proxy_config:go_default_library",
         "//internal/platform/postgres_config/client_config:go_default_library",
+        "//internal/platform/postgres_config/feed/actor_profile_record_config:go_default_library",
     ],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["add_proxy_test.go"],
-    embed = [":go_default_library"],
-    deps = ["//vendor/github.com/stretchr/testify/assert:go_default_library"],
 )
