+++ cmd/lambda/TCR/add_proxy/main/BUILD.bazel
@@ -1,4 +1,4 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
@@ -9,10 +9,4 @@ go_library(
         "//cmd/lambda/TCR/add_proxy/config:go_default_library",
         "//vendor/github.com/aws/aws-lambda-go/lambda:go_default_library",
     ],
-)
-
-go_binary(
-    name = "main",
-    embed = [":go_default_library"],
-    visibility = ["//visibility:public"],
 )
