+++ cmd/lambda/TCR/add_proxy/main/main.go
@@ -6,5 +6,5 @@ import (
 )
 
 func main() {
-  lambda.Start(config.Handler)
+  lambda.Start(lambda_add_proxy_config.Handler)
 }
