+++ cmd/lambda/TCR/add_proxy_voting_for_principal/config/BUILD.bazel
@@ -1,0 +1,16 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["add_proxy_voting_for_principal.go"],
+    importpath = "BigBang/cmd/lambda/TCR/add_proxy_voting_for_principal/config",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//internal/app/tcr_attributes:go_default_library",
+        "//internal/pkg/error_config:go_default_library",
+        "//internal/platform/postgres_config/TCR/principal_proxy_votes_config:go_default_library",
+        "//internal/platform/postgres_config/TCR/proxy_config:go_default_library",
+        "//internal/platform/postgres_config/client_config:go_default_library",
+        "//internal/platform/postgres_config/feed/actor_profile_record_config:go_default_library",
+    ],
+)
