+++ cmd/lambda/TCR/add_proxy_voting_for_principal/main/BUILD.bazel
@@ -1,0 +1,18 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["main.go"],
+    importpath = "BigBang/cmd/lambda/TCR/add_proxy_voting_for_principal/main",
+    visibility = ["//visibility:private"],
+    deps = [
+        "//cmd/lambda/TCR/add_proxy_voting_for_principal/config:go_default_library",
+        "//vendor/github.com/aws/aws-lambda-go/lambda:go_default_library",
+    ],
+)
+
+go_binary(
+    name = "main",
+    embed = [":go_default_library"],
+    visibility = ["//visibility:public"],
+)
