--- cmd/lambda/TCR/get_proxy_votes_list_by_principal/config/BUILD.bazel
@@ -1,18 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["get_proxy_votes_list_by_principal.go"],
-    importpath = "BigBang/cmd/lambda/TCR/get_proxy_votes_list_by_principal/config",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//internal/app/tcr_attributes:go_default_library",
-        "//internal/pkg/error_config:go_default_library",
-        "//internal/pkg/utils:go_default_library",
-        "//internal/platform/postgres_config/TCR/actor_delegate_votes_account_config:go_default_library",
-        "//internal/platform/postgres_config/TCR/principal_proxy_votes_config:go_default_library",
-        "//internal/platform/postgres_config/TCR/project_config:go_default_library",
-        "//internal/platform/postgres_config/client_config:go_default_library",
-        "//internal/platform/postgres_config/feed/actor_profile_record_config:go_default_library",
-    ],
-)
