--- cmd/lambda/TCR/get_proxy_votes_list_by_principal/main/BUILD.bazel
@@ -1,18 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["main.go"],
-    importpath = "BigBang/cmd/lambda/TCR/get_proxy_votes_list_by_principal/main",
-    visibility = ["//visibility:private"],
-    deps = [
-        "//cmd/lambda/TCR/get_proxy_votes_list_by_principal/config:go_default_library",
-        "//vendor/github.com/aws/aws-lambda-go/lambda:go_default_library",
-    ],
-)
-
-go_binary(
-    name = "main",
-    embed = [":go_default_library"],
-    visibility = ["//visibility:public"],
-)
