+++ scripts/lambda/generate_all_feed_lambda_zips_for_all_stages.sh
@@ -1,0 +1,10 @@
+#!/usr/bin/env bash
+
+echo "generating test"
+.//generate_all_feed_lambda_zips.sh v3 test
+
+echo "generating staging"
+.//generate_all_feed_lambda_zips.sh v3 staging
+
+echo "generating exp"
+.//generate_all_feed_lambda_zips.sh v3 exp
