+++ scripts/lambda/update_all_tcr_lambda_zips_for_all_stages.sh
@@ -1,0 +1,10 @@
+#!/usr/bin/env bash
+
+echo "updating test"
+.//update_all_tcr_lambda_zips.sh v3  test
+
+echo "updating staging"
+.//update_all_tcr_lambda_zips.sh v3 staging
+
+echo "updating exp"
+.//update_all_tcr_lambda_zips.sh v3 exp
