--- scripts/lambda/update_lambda_zip_for_all_stages.sh
@@ -1,10 +1,0 @@
-#!/usr/bin/env bash
-
-echo "updating testing"
-.//update_lambda_zip.sh $1 $2 test
-
-echo "updating staging"
-.//update_lambda_zip.sh $1 $2 staging
-
-echo "updating production"
-.//update_lambda_zip.sh $1 $2 producction
