+++ scripts/test/run_all_lambda_tcr_tests.sh
@@ -18,16 +18,18 @@ bazel run //:gazelle
 
 
 ./run_unit_test.sh  //test/lambda/TCR/rating_vote:go_default_test
+
+./run_unit_test.sh  //test/lambda/TCR/add_proxy:go_default_test
 ./run_unit_test.sh  //test/lambda/TCR/update_delegate_votes:go_default_test
-./run_unit_test.sh  //test/lambda/TCR/adjust_proxy_votes:go_default_test
+./run_unit_test.sh  //test/lambda/TCR/add_proxy_voting_for_principal:go_default_test
 
 ./run_unit_test.sh  //test/lambda/TCR/get_rating_vote_list:go_default_test
 ./run_unit_test.sh  //test/lambda/TCR/get_proxy_voting_info:go_default_test
 
-./run_unit_test.sh  //test/lambda/TCR/delete_objective:go_default_test
-./run_unit_test.sh  //test/lambda/TCR/delete_milestone:go_default_test
-./run_unit_test.sh  //test/lambda/TCR/delete_project:go_default_test
-
-#./run_unit_test.sh  //cmd/lambda/TCR/add_proxy/config:go_default_test
+#./run_unit_test.sh  //test/lambda/TCR/delete_objective:go_default_test
+#./run_unit_test.sh  //test/lambda/TCR/delete_milestone:go_default_test
+#./run_unit_test.sh  //test/lambda/TCR/delete_project:go_default_test
+#
+#
 #./run_unit_test.sh  //cmd/lambda/TCR/get_proxy_list/config:go_default_test
 #./run_unit_test.sh  //cmd/lambda/TCR/delete_proxy/config:go_default_test
