+++ test/lambda/TCR/add_proxy_voting_for_principal/BUILD.bazel
@@ -1,0 +1,12 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_test")
+
+go_test(
+    name = "go_default_test",
+    srcs = ["add_proxy_voting_for_principal_test.go"],
+    deps = [
+        "//cmd/lambda/TCR/add_proxy_voting_for_principal/config:go_default_library",
+        "//internal/app/tcr_attributes:go_default_library",
+        "//test/constants:go_default_library",
+        "//vendor/github.com/stretchr/testify/assert:go_default_library",
+    ],
+)
