--- test/lambda/TCR/adjust_proxy_votes/BUILD.bazel
@@ -1,11 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_test")
-
-go_test(
-    name = "go_default_test",
-    srcs = ["adjust_proxy_votes_test.go"],
-    deps = [
-        "//cmd/lambda/TCR/adjust_proxy_votes/config:go_default_library",
-        "//test/constants:go_default_library",
-        "//vendor/github.com/stretchr/testify/assert:go_default_library",
-    ],
-)
