+++ test/lambda/TCR/get_proxy_voting_info/BUILD.bazel
@@ -4,7 +4,7 @@ go_test(
     name = "go_default_test",
     srcs = ["get_proxy_voting_info_test.go"],
     deps = [
-        "//cmd/lambda/TCR/get_proxy_votes_list_by_principal/config:go_default_library",
+        "//cmd/lambda/TCR/get_proxy_voting_info/config:go_default_library",
         "//internal/app/tcr_attributes:go_default_library",
         "//internal/platform/postgres_config/TCR/principal_proxy_votes_config:go_default_library",
         "//test/constants:go_default_library",
